requirements.txt
mysql.connector
validate_email
bcrypt